% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doughut.R
\name{create_rings}
\alias{create_rings}
\title{Make doughnut}
\usage{
create_rings(point, n_circles, distance = 1)
}
\arguments{
\item{point}{Optional midpoint of the region}

\item{n_circles}{Number of rings including the central circle}

\item{distance}{Distance between each doughnut ring in km}

\item{n}{Number of doughnuts}

\item{intersection}{Intersection with x?}
}
\value{
An sf object containing concentric donuts
}
\description{
Make doughnut
}
\examples{
x = zb_region
plot(zb_doughnut(x, n = 4))
plot(zb_doughnut(x, d = 4))
x_point = sf::st_centroid(zb_region)
zb_doughnut(point = x_point, n = 4)
}
