% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{london_area}
\alias{london_area}
\alias{london_cent}
\title{Region representing London in projected coordinate system}
\description{
`london_area` and `london_cent` represent the city boundaries and centre point of London, respectively.
}
\note{
`london_area` is a projected version of the `lnd` object in the `spDataLarge` package.
See the `data-raw` folder in the package's repo to reproduce these datasets
}
\examples{
plot(london_area, reset = FALSE)
plot(london_cent, add = TRUE)
}
\keyword{datasets}
