% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sz_lines.R
\name{sz_lines}
\alias{sz_lines}
\title{Create lines radiating at equal angles from a point}
\usage{
sz_lines(point, n, starting_angle = 45, distance = 1e+05)
}
\arguments{
\item{point}{}

\item{n}{}

\item{starting_angle}{}

\item{distance}{}
}
\value{

}
\description{
Create lines radiating at equal angles from a point
}
\examples{
point = sf::st_centroid(sz_region)
n = 4
}
