% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{sz_segment}
\alias{sz_segment}
\title{Title}
\usage{
sz_segment(x, ncol, nrow = NULL, intersection = TRUE)
}
\arguments{
\item{point}{}

\item{n}{}
}
\value{

}
\description{
Title
}
\examples{
point = sf::st_centroid(sz_region)
n = 4
segments = sz_segment(point, n)
}
