% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{sz_zone}
\alias{sz_zone}
\title{Generate zones covering a study region}
\usage{
sz_zone(
  x,
  n_circles,
  n_segments = c(1, (1:(n_circles - 1)) * 4),
  intersection = TRUE
)
}
\arguments{
\item{n_circles}{Number of rings including the central circle}

\item{n_segments}{Optional sequence of numbers}

\item{point}{Optional midpoint of the region}
}
\value{

}
\description{
This function first divides geographic space into [annuli](https://en.wikipedia.org/wiki/Annulus_(mathematics)) 
(concentric 2d rings or 'dohnuts') and then subdivides each annulus
into a number of segments.
}
\examples{
z = sz_zone(sz_region, n_circles = 4)
z
plot(z, col = 1:nrow(z))
z = sz_zone(sz_region, n_circles = 6)
plot(z, col = 1:nrow(z))
}
