% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zb_100_triangular_numbers}
\alias{zb_100_triangular_numbers}
\title{The first 100 triangular numbers}
\description{
The first 100 in the sequence of [triangular numbers](https://en.wikipedia.org/wiki/Triangular_number)
}
\note{
See the `data-raw` folder in the package's repo to reproduce these datasets
}
\keyword{datasets}
