% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doughut.R
\name{zb_doughnut}
\alias{zb_doughnut}
\title{Make doughnuts}
\usage{
zb_doughnut(x = NULL, point = NULL, n_circles = NULL,
  distance = NULL, distance_growth = 1)
}
\arguments{
\item{x}{Region of interest}

\item{point}{Optional midpoint of the region}

\item{n_circles}{Number of rings including the central circle}

\item{distance}{Distance  (km)}

\item{distance_growth}{The rate at which the doughnut ring widths grow (km)}
}
\value{
An sf object containing concentric donuts
}
\description{
Make doughnuts
}
\examples{
x = zb_region
plot(zb_doughnut(x, n_circles = 4), reset = FALSE)
plot(zb_region, add = TRUE)
z = zb_doughnut(x, distance = 0.5)
plot(z, reset = FALSE)
plot(zb_region, add = TRUE)
plot(z[5, ], add = TRUE, col = "red")
x_point = sf::st_centroid(zb_region)
z = zb_doughnut(point = zb_region_cent, n_circles = 4, distance = 4:1)
z
plot(z)
}
