% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_lines.R
\name{zb_lines}
\alias{zb_lines}
\title{Create lines radiating at equal angles from a point}
\usage{
zb_lines(point, n, starting_angle = 45, distance = 1e+05)
}
\arguments{
\item{point}{Center point}

\item{n}{Number of lines}

\item{starting_angle}{Starting angle}

\item{distance}{Distance}
}
\value{
Objects of class `sfc` containing linestring geometries
}
\description{
Create lines radiating at equal angles from a point
}
\examples{
point = sf::st_centroid(london_a())
n = 4
l = zb_lines(point, n)
plot(l)
}
