% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{zb_palette}
\alias{zb_palette}
\title{Generate a colour palette for zones}
\usage{
zb_palette(z)
}
\arguments{
\item{z}{An `sf` object containing zones covering the region}
}
\value{
A vector of colours
}
\description{
This function generates a colour palette for zones, using hcl colour space
}
\examples{
z = zb_zone(zb_region, point = zb_region_cent)
zb_palette(z)
plot(z, col = zb_palette(z))
}
