% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrat.R
\name{zb_quadrat}
\alias{zb_quadrat}
\title{Divide a region into quadrats}
\usage{
zb_quadrat(x, ncol, nrow = NULL, intersection = TRUE)
}
\arguments{
\item{x}{x}

\item{ncol}{ncol}

\item{nrow}{nrow}

\item{intersection}{intersection}
}
\value{
An sf object
}
\description{
Divide a region into quadrats
}
\examples{
x = london_a()
c = sf::st_centroid(london_a())
plot(zb_quadrat(x, ncol = 2), col = 2:5)
plot(c, add = TRUE, col = "white")
plot(zb_quadrat(x, ncol = 3))
plot(zb_quadrat(x, ncol = 4))
plot(zb_quadrat(x, ncol = 4, intersection = FALSE))
}
