% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zb_region}
\alias{zb_region}
\alias{zb_region_cent}
\title{Region representing London in projected coordinate system}
\description{
`zb_region` and `zb_region_cent` represent the city boundaries and centre point of London, respectively.
}
\note{
`zb_region` is a projected version of the `lnd` object in the `spDataLarge` package.
See the `data-raw` folder in the package's repo to reproduce these datasets
}
\examples{
plot(zb_region, reset = FALSE)
plot(zb_region_cent$geometry, add = TRUE)
}
\keyword{datasets}
