% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{zb_segment}
\alias{zb_segment}
\title{Title}
\usage{
zb_segment(x = NULL, area = NULL, n_segments = 12, distance = NA)
}
\arguments{
\item{x}{point}

\item{area}{(optional) Area. Should be an \code{\link[sf:sf]{sf}} or \code{\link[sf:sfc]{sfc}} object containing one (multi) polygon}

\item{n_segments}{Number of segments}

\item{distance}{Distance in metres}

\item{starting_angle}{Starting angle}
}
\value{
An sf object with polygons representing segments dividing up geographic space
}
\description{
Title
}
\examples{
point = sf::st_centroid(london_area)
n = 4
segments = zb_segment(point, n)
plot(segments, col = 1:n)
n = 16
segments = zb_segment(point, n)
plot(segments, col = 1:n) # logo?
}
