% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{zb_view}
\alias{zb_view}
\title{View zones}
\usage{
zb_view(z)
}
\arguments{
\item{z}{An `sf` object containing zones covering the region}
}
\description{
This function opens an interactive map of the zones
}
