% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{zb_zone}
\alias{zb_zone}
\title{Generate zones covering a region of interest}
\usage{
zb_zone(
  x = NULL,
  point = NULL,
  n_circles = NULL,
  n_segments = 12,
  distance = 1,
  distance_growth = 1,
  starting_angle = 0,
  segment_center = FALSE,
  intersection = TRUE
)
}
\arguments{
\item{x}{Region of interest}

\item{point}{Optional midpoint of the region}

\item{n_circles}{Number of rings including the central circle}

\item{n_segments}{Optional sequence of numbers}

\item{distance}{Distance  (km)}

\item{distance_growth}{The rate at which the doughnut ring widths grow (km)}

\item{starting_angle}{The angle of the first of the radii that create the segments (degrees)}

\item{segment_center}{Should the central circle be divided into segments? `FALSE` by default.}

\item{intersection}{Not implemented yet}
}
\value{
An `sf` object containing zones covering the region
}
\description{
This function first divides geographic space into [annuli](https://en.wikipedia.org/wiki/Annulus_(mathematics)) 
(concentric 2d rings or 'doughnuts') and then subdivides each annulus
into a number of segments.
}
\examples{
z = zb_zone(zb_region, n_circles = 4)
z
plot(z)
plot(zb_zone(zb_region, point = zb_region_cent))
plot(zb_zone(zb_region, n_circles = 6))
plot(zb_zone(zb_region, starting_angle = -15))
plot(zb_zone(zb_region, n_circles = 8, distance = 0.1, distance_growth = 0.1))
if (require(tmap)) {
  z = zb_zone(zb_region, n_circles = 8)
  tmap_mode("view")
  tm_shape(z) + tm_polygons("circle_id", palette = "plasma", legend.show = FALSE) + tm_text("label")
  z = zb_zone(zb_region, n_circles = 8, n_segments = c(1, 4, 4, 8, 8, 16, 16, 16))
  tm_shape(z) + tm_polygons("circle_id", palette = "plasma", legend.show = FALSE) + tm_text("label")
}
}
