% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{london_area}
\alias{london_area}
\alias{london_a}
\alias{london_cent}
\alias{london_c}
\alias{london_cent_lonlat}
\alias{london_area_lonlat}
\title{Region representing London in projected coordinate system}
\usage{
london_a()

london_c()
}
\description{
`london_a()` and `london_c()` return the city boundaries and centre
  point of London, respectively.
}
\note{
`london_a()`  returns a projected version of `lnd` in `spDataLarge`.
See the `data-raw` folder in the package's repo to reproduce these datasets
The `lonlat` versions of the data have coordinates in units of degrees.
}
\examples{
plot(london_a(), reset = FALSE)
plot(london_c(), add = TRUE)
}
\keyword{datasets}
