% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doughnut.R
\name{zb_doughnut}
\alias{zb_doughnut}
\title{Title}
\usage{
zb_doughnut(x = NULL, point = NULL, n = NULL, distance = 1, intersection = TRUE)
}
\arguments{
\item{x}{x}

\item{point}{Optional midpoint of the region}

\item{n}{Number of doughnuts}

\item{distance}{Distance between each doughnut ring in km}

\item{intersection}{Intersection with x?}
}
\value{
An sf object containing concentric donuts
}
\description{
Title
}
\examples{
x = zb_region
plot(zb_doughnut(x, n = 4))
plot(zb_doughnut(x, d = 4))
}
