% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dohnut.R
\name{sz_dohnut}
\alias{sz_dohnut}
\title{Title}
\usage{
sz_dohnut(x, c = sf::st_centroid(x), n = NULL, d = 1,
  intersection = TRUE)
}
\arguments{
\item{x}{Spatial object}

\item{c}{Center point}

\item{n}{Number of dohnuts}

\item{d}{Distance between each dohnut ring in km}

\item{intersection}{Intersection with x?}
}
\description{
Title
}
\examples{
x = sz_region
plot(sz_dohnut(x, n = 4))
plot(sz_dohnut(x, d = 4))
}
