% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dohnut.R
\name{sz_dohnut}
\alias{sz_dohnut}
\title{Title}
\usage{
sz_dohnut(x = NULL, point = NULL, n = NULL, distance = 1, intersection = TRUE)
}
\arguments{
\item{x}{x}

\item{point}{Optional midpoint of the region}

\item{n}{Number of dohnuts}

\item{distance}{Distance between each dohnut ring in km}

\item{intersection}{Intersection with x?}
}
\value{
An sf object containing concentric donuts
}
\description{
Title
}
\examples{
x = sz_region
plot(sz_dohnut(x, n = 4))
plot(sz_dohnut(x, d = 4))
}
