% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{sz_segment}
\alias{sz_segment}
\title{Title}
\usage{
sz_segment(x, n = 4, starting_angle = 45, distance = 1e+05)
}
\arguments{
\item{x}{}

\item{n}{Number of segmentss}

\item{starting_angle}{Starting angle}

\item{distance}{distance}
}
\description{
Title
}
\examples{
point = sf::st_centroid(sz_region)
n = 4
segments = sz_segment(point, n)
plot(segments, col = 1:n)
n = 16
segments = sz_segment(point, n)
plot(segments, col = 1:n) # logo?
}
