% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{zb_segment}
\alias{zb_segment}
\title{Title}
\usage{
zb_segment(x, n_segments = 4, starting_angle = 45, distance = 1e+05)
}
\arguments{
\item{x}{Region of interest}

\item{n_segments}{Number of segments}

\item{starting_angle}{Starting angle}

\item{distance}{Distance in metres}
}
\value{
An sf object with polygons representing segments dividing up geographic space
}
\description{
Title
}
\examples{
point = sf::st_centroid(zb_region)
n = 4
segments = zb_segment(point, n)
plot(segments, col = 1:n)
n = 16
segments = zb_segment(point, n)
plot(segments, col = 1:n) # logo?
}
