% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zb_region}
\alias{zb_region}
\title{Region representing London in projected coordinate system}
\description{
Region representing London in projected coordinate system
}
\note{
This is a projected version of the `lnd` object in the `spDataLarge` package
}
\keyword{datasets}
