% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{zb_zone}
\alias{zb_zone}
\title{Generate zones covering a region of interest}
\usage{
zb_zone(
  x = NULL,
  point = NULL,
  n_circles = NULL,
  n_segments = 12,
  distance = 1,
  distance_growth = 1,
  intersection = TRUE
)
}
\arguments{
\item{x}{Region of interest}

\item{point}{Optional midpoint of the region}

\item{n_circles}{Number of rings including the central circle}

\item{n_segments}{Optional sequence of numbers}

\item{distance}{Distance  (km)}

\item{distance_growth}{The rate at which the doughnut ring widths grow (km)}

\item{intersection}{Not implemented yet}
}
\value{
An `sf` object containing zones covering the region
}
\description{
This function first divides geographic space into [annuli](https://en.wikipedia.org/wiki/Annulus_(mathematics)) 
(concentric 2d rings or 'doughnuts') and then subdivides each annulus
into a number of segments.
}
\examples{
data(zb_region)
z = zb_zone(zb_region, n_circles = 4)
z
plot(z, col = 1:nrow(z))
z_from_cent = zb_zone(zb_region, point = zb_region_cent, n_circles = 8)
plot(z_from_cent, col = 1:nrow(z))
zb_region_sf = sf::st_sf(data.frame(n = 1), geometry = zb_region)
z = zb_zone(zb_region_sf, n_circles = 3, n_segments = c(1, 4, 8))
plot(z) # quadrant not respected 
plot(zb_zone(zb_region, n_circles = 6), col = 1:6)
plot(zb_zone(zb_region, n_circles = 8), col = 1:8)
plot(zb_zone(zb_region, n_circles = 8, distance = 0.1), col = 1:8)
}
