% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{zb_zone}
\alias{zb_zone}
\title{Generate zones covering a region of interest}
\usage{
zb_zone(
  x = NULL,
  point = NULL,
  n_circles = NULL,
  n_segments = 12,
  distance = 1,
  distance_growth = 1,
  starting_angle = 15,
  segment_center = FALSE,
  intersection = TRUE
)
}
\arguments{
\item{x}{Region of interest}

\item{point}{Optional midpoint of the region}

\item{n_circles}{Number of rings including the central circle}

\item{n_segments}{Optional sequence of numbers}

\item{distance}{Distance  (km)}

\item{distance_growth}{The rate at which the doughnut ring widths grow (km)}

\item{starting_angle}{The angle of the first of the radii that create the segments (degrees)}

\item{segment_center}{Should the central circle be divided into segments? `FALSE` by default.}

\item{intersection}{Not implemented yet}
}
\value{
An `sf` object containing zones covering the region
}
\description{
This function first divides geographic space into [annuli](https://en.wikipedia.org/wiki/Annulus_(mathematics)) 
(concentric 2d rings or 'doughnuts') and then subdivides each annulus
into a number of segments.
}
\details{
By default 12 segments are used for each annuli, resulting in a zoning system that can 
be used to refer to segments in [clock position](https://en.wikipedia.org/wiki/Clock_position),
with 12 representing North, 3 representing East, 6 Sounth and 9 Western segments.
}
\examples{
z = zb_zone(london_area, point = london_cent)
z
plot(z)
plot(zb_zone(london_area, n_circles = 2))
plot(zb_zone(london_area, n_circles = 2, starting_angle = 0))
plot(zb_zone(london_area,n_circles = 2, starting_angle = 0, distance_growth = 0.1))

if (require(tmap)) {
  # tmap_mode("view") # for interactive maps
  z = zb_zone(london_area, point = london_cent)
  tm_shape(z) + tm_polygons("circle_id", palette = "plasma", legend.show = FALSE) + tm_text("label")
}
}
